fix: Make MediaGallery drag reorg work
