+++ .circleci/tests.sh
@@ -6,8 +6,6 @@
 # often avoided and the false positive doesn't fail the CircleCI build.
 #
 # This will update npm deps which takes a while
-# cfs:tempstore: updating npm dependencies -- combined-stream...
-# cfs:gridfs: updating npm dependencies -- mongodb, gridfs-stream...
 
 set +e
