+++ imports/plugins/core/ui/client/containers/tagListContainer.js
@@ -3,13 +3,12 @@ import PropTypes from "prop-types";
 import _ from "lodash";
 import update from "immutability-helper";
 import { compose } from "recompose";
-import { registerComponent, composeWithTracker } from "@reactioncommerce/reaction-components";
+import { registerComponent, composeWithTracker, Components } from "@reactioncommerce/reaction-components";
 import { Meteor } from "meteor/meteor";
 import { Reaction, i18next } from "/client/api";
 import TagList from "../components/tags/tagList";
 import { Tags } from "/lib/collections";
 import { getTagIds } from "/lib/selectors/tags";
-import { DragDropProvider } from "/imports/plugins/core/ui/client/providers";
 
 function updateSuggestions(term, { excludeTags }) {
   const slug = Reaction.getSlug(term);
@@ -214,7 +213,7 @@ const wrapComponent = (Comp) => (
 
     render() {
       return (
-        <DragDropProvider>
+        <Components.DragDropProvider>
           <Comp
             newTag={this.state.newTag}
             onClick={this.handleEditButtonClick}
@@ -231,7 +230,7 @@ const wrapComponent = (Comp) => (
             tooltip="Unpublished changes"
             {...this.props}
           />
-        </DragDropProvider>
+        </Components.DragDropProvider>
       );
     }
   }
