+++ imports/plugins/included/product-detail-simple/client/containers/variantList.js
@@ -9,7 +9,6 @@ import { getChildVariants } from "../selectors/variants";
 import { Products } from "/lib/collections";
 import update from "immutability-helper";
 import { getVariantIds } from "/lib/selectors/variants";
-import { DragDropProvider } from "/imports/plugins/core/ui/client/providers";
 import { Media } from "/imports/plugins/core/files/client";
 
 function variantIsSelected(variantId) {
@@ -160,7 +159,7 @@ class VariantListContainer extends Component {
 
   render() {
     return (
-      <DragDropProvider>
+      <Components.DragDropProvider>
         <Components.VariantList
           onEditVariant={this.handleEditVariant}
           onMoveVariant={this.handleMoveVariant}
@@ -170,7 +169,7 @@ class VariantListContainer extends Component {
           {...this.props}
           variants={this.variants}
         />
-      </DragDropProvider>
+      </Components.DragDropProvider>
     );
   }
 }
