+++ imports/plugins/included/product-variant/containers/productGridContainer.js
@@ -100,7 +100,7 @@ const wrapComponent = (Comp) => (
       const dragProductId = this.state.productIds[dragIndex];
       const hoverProductId = this.state.productIds[hoverIndex];
       const dragProductWeight = _.get(this, `state.productsByKey[${dragProductId}].positions[${tag}].weight`, 0);
-      const dropProductWeight = _.get(this ,`state.productsByKey[${hoverProductId}].positions[${tag}].weight`, 0);
+      const dropProductWeight = _.get(this, `state.productsByKey[${hoverProductId}].positions[${tag}].weight`, 0);
 
       const newState = update(this.state, {
         productsByKey: {
